requirements.txt
Django~=3.2.10
djangorestframework==3.13.1
django-filter==22.1